Switching to cleaner method of canonicalizing code coverage paths
The previous method used hacky regexes to tamper with the .coverage file. This method uses supported features of the coverage binary to create a coverage file within the Docker container, then use "coverage combine" outside the container so that coverage canonicalizes all of the source file paths to match how they look on the local filesystem.